requirements.txt
hypercorn
epitran
textblob
translate
gtts
pygame
mysql-connector-python
python-dotenv
uvicorn
pyyaml
python-multipart